moved helper classes to anonymous inner classes in runner initializers; log all exceptions
