+++ web/web-base/src/main/java/org/visallo/web/initializers/GraphPropertyWorkerRunnerInitializer.java
@@ -1,24 +1,27 @@
 package org.visallo.web.initializers;
 
 import com.google.inject.Inject;
+import org.visallo.core.bootstrap.InjectHelper;
 import org.visallo.core.config.Configuration;
+import org.visallo.core.ingest.graphProperty.GraphPropertyRunner;
 import org.visallo.core.model.user.UserRepository;
 import org.visallo.core.user.User;
+import org.visallo.core.util.StoppableRunnable;
 import org.visallo.core.util.VisalloLogger;
 import org.visallo.core.util.VisalloLoggerFactory;
 
 import javax.servlet.ServletContext;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
 public class GraphPropertyWorkerRunnerInitializer extends ApplicationBootstrapInitializer {
-    public static final String CONFIG_THREAD_COUNT = GraphPropertyWorkerRunnerInitializer.class.getName() + ".threadCount";
-    public static final int DEFAULT_THREAD_COUNT = 1;
+    private static final String CONFIG_THREAD_COUNT = GraphPropertyWorkerRunnerInitializer.class.getName() + ".threadCount";
+    private static final int DEFAULT_THREAD_COUNT = 1;
     private static final VisalloLogger LOGGER = VisalloLoggerFactory.getLogger(GraphPropertyWorkerRunnerInitializer.class);
+
     private final Configuration config;
     private final UserRepository userRepository;
-    private List<GraphPropertyWorkerRunnerHelper> workerHelpers = new ArrayList<>();
+    private final List<StoppableRunnable> stoppables = new ArrayList<>();
 
     @Inject
     public GraphPropertyWorkerRunnerInitializer(
@@ -34,13 +37,37 @@ public class GraphPropertyWorkerRunnerInitializer extends ApplicationBootstrapIn
         LOGGER.debug("setupGraphPropertyWorkerRunner");
 
         int threadCount = config.getInt(CONFIG_THREAD_COUNT, DEFAULT_THREAD_COUNT);
-        final User user = userRepository.getSystemUser();
+        User user = userRepository.getSystemUser();
 
         LOGGER.debug("starting graph property worker runners: %d", threadCount);
         for (int i = 0; i < threadCount; i++) {
-            GraphPropertyWorkerRunnerHelper workerHelper = new GraphPropertyWorkerRunnerHelper(user);
-            workerHelpers.add(workerHelper);
-            Thread t = new Thread(workerHelper);
+            StoppableRunnable stoppable = new StoppableRunnable() {
+                private GraphPropertyRunner graphPropertyRunner = null;
+
+                @Override
+                public void run() {
+                    try {
+                        graphPropertyRunner = InjectHelper.getInstance(GraphPropertyRunner.class);
+                        graphPropertyRunner.prepare(user);
+                        graphPropertyRunner.run();
+                    } catch (Exception ex) {
+                        LOGGER.error("Failed running GraphPropertyRunner", ex);
+                    }
+                }
+
+                @Override
+                public void stop() {
+                    try {
+                        if (graphPropertyRunner != null) {
+                            graphPropertyRunner.stop();
+                        }
+                    } catch (Exception ex) {
+                        LOGGER.error("Failed stopping GraphPropertyRunner", ex);
+                    }
+                }
+            };
+            stoppables.add(stoppable);
+            Thread t = new Thread(stoppable);
             t.setName("graph-property-worker-runner-" + t.getId());
             t.setDaemon(true);
             LOGGER.debug("starting graph property worker runner thread: %s", t.getName());
@@ -49,9 +76,7 @@ public class GraphPropertyWorkerRunnerInitializer extends ApplicationBootstrapIn
     }
 
     @Override
-    public void close() throws IOException {
-        for (GraphPropertyWorkerRunnerHelper runner : workerHelpers) {
-            runner.stop();
-        }
+    public void close() {
+        stoppables.forEach(StoppableRunnable::stop);
     }
 }
