+++ web/web-base/src/main/java/org/visallo/web/initializers/LongRunningProcessRunnerInitializer.java
@@ -1,21 +1,24 @@
 package org.visallo.web.initializers;
 
 import com.google.inject.Inject;
+import org.visallo.core.bootstrap.InjectHelper;
 import org.visallo.core.config.Configuration;
+import org.visallo.core.model.longRunningProcess.LongRunningProcessRunner;
+import org.visallo.core.util.StoppableRunnable;
 import org.visallo.core.util.VisalloLogger;
 import org.visallo.core.util.VisalloLoggerFactory;
 
 import javax.servlet.ServletContext;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
 public class LongRunningProcessRunnerInitializer extends ApplicationBootstrapInitializer {
-    public static final String CONFIG_THREAD_COUNT = LongRunningProcessRunnerInitializer.class.getName() + ".threadCount";
-    public static final int DEFAULT_THREAD_COUNT = 1;
+    private static final String CONFIG_THREAD_COUNT = LongRunningProcessRunnerInitializer.class.getName() + ".threadCount";
+    private static final int DEFAULT_THREAD_COUNT = 1;
     private static final VisalloLogger LOGGER = VisalloLoggerFactory.getLogger(LongRunningProcessRunnerInitializer.class);
+
     private final Configuration config;
-    private List<LongRunningProcessRunnerHelper> runnerHelpers = new ArrayList<>();
+    private final List<StoppableRunnable> stoppables = new ArrayList<>();
 
     @Inject
     public LongRunningProcessRunnerInitializer(Configuration config) {
@@ -30,9 +33,33 @@ public class LongRunningProcessRunnerInitializer extends ApplicationBootstrapIni
 
         LOGGER.debug("long running process runners: %d", threadCount);
         for (int i = 0; i < threadCount; i++) {
-            LongRunningProcessRunnerHelper runnerHelper = new LongRunningProcessRunnerHelper(config);
-            runnerHelpers.add(runnerHelper);
-            Thread t = new Thread(runnerHelper);
+            StoppableRunnable stoppable = new StoppableRunnable() {
+                private LongRunningProcessRunner longRunningProcessRunner = null;
+
+                @Override
+                public void run() {
+                    try {
+                        longRunningProcessRunner = InjectHelper.getInstance(LongRunningProcessRunner.class);
+                        longRunningProcessRunner.prepare(config.toMap());
+                        longRunningProcessRunner.run();
+                    } catch (Exception ex) {
+                        LOGGER.error("Failed running LongRunningProcessRunner", ex);
+                    }
+                }
+
+                @Override
+                public void stop() {
+                    try {
+                        if (longRunningProcessRunner != null){
+                            longRunningProcessRunner.stop();
+                        }
+                    } catch (Exception ex) {
+                        LOGGER.error("Failed stopping LongRunningProcessRunner", ex);
+                    }
+                }
+            };
+            stoppables.add(stoppable);
+            Thread t = new Thread(stoppable);
             t.setName("long-running-process-runner-" + t.getId());
             t.setDaemon(true);
             LOGGER.debug("starting long running process runner thread: %s", t.getName());
@@ -41,9 +68,7 @@ public class LongRunningProcessRunnerInitializer extends ApplicationBootstrapIni
     }
 
     @Override
-    public void close() throws IOException {
-        for (LongRunningProcessRunnerHelper runnerHelper : runnerHelpers) {
-            runnerHelper.stop();
-        }
+    public void close() {
+        stoppables.forEach(StoppableRunnable::stop);
     }
 }
